feat(frontend): merge ThemeTab and TextTab into AppearanceTab
Consolidate separate Theme and Text settings tabs into a unified
Appearance tab for a cleaner settings panel layout.

- Create AppearanceTab combining theme and text settings
- Remove standalone ThemeTab and TextTab components
- Update SettingsPanel, GeneralTab, and AboutTab accordingly
- Rename bookmarks section labels for clarity
