+++ frontend/src/composables/useI18n.ts
@@ -62,8 +62,8 @@ const messages: Record<Locale, Record<string, string>> = {
     'terminal.ctxCommand': 'Command',
     'terminal.ctxSave': 'Save',
     'terminal.ctxNoSelection': 'No text selected',
-    'bookmarks.title': 'Saved Commands',
-    'bookmarks.empty': 'No saved commands yet',
+    'bookmarks.title': 'Bookmarks',
+    'bookmarks.empty': 'No bookmarks yet',
     'bookmarks.name': 'Name',
     'bookmarks.command': 'Command',
     'bookmarks.group': 'Group',
@@ -92,8 +92,7 @@ const messages: Record<Locale, Record<string, string>> = {
     'settings.color.ansi': 'ANSI Colors',
     'settings.color.reset': 'Reset',
     'settings.tab.general': 'General',
-    'settings.tab.theme': 'Theme',
-    'settings.tab.text': 'Text',
+    'settings.tab.appearance': 'Appearance',
     'settings.tab.keyboard': 'Keyboard',
     'settings.text': 'Text Settings',
     'settings.text.fontSize': 'Font Size',
@@ -216,9 +215,13 @@ const messages: Record<Locale, Record<string, string>> = {
     'settings.tab.plugins': 'Plugins',
     'settings.plugins.install': 'Install Plugin',
     'settings.plugins.installFile': 'Install from File',
+    'settings.plugins.installFolder': 'Install from Folder',
     'settings.plugins.installed': 'Installed Plugins',
     'settings.plugins.devLink': 'Development',
     'settings.plugins.devLinkHint': 'Symlink a local plugin directory for hot-reload development',
+    'settings.plugins.devLinkCheckbox': 'Dev Link (hot-reload)',
+    'settings.plugins.browseFolder': 'Browse...',
+    'settings.plugins.devBadge': 'dev',
     'settings.plugins.load': 'Load',
     'settings.plugins.update': 'Update',
     'settings.plugins.uninstall': 'Uninstall',
@@ -240,6 +243,7 @@ const messages: Record<Locale, Record<string, string>> = {
     'settings.plugins.noReadme': 'No detailed description available',
     'settings.plugins.back': 'Back',
     'settings.plugins.installedBadge': 'Installed',
+    'settings.plugins.refresh': 'Refresh',
     'login.subtitle': 'Enter your access token to continue',
     'login.placeholder': 'Access Token',
     'login.submit': 'Sign In',
@@ -255,7 +259,7 @@ const messages: Record<Locale, Record<string, string>> = {
     'setup.saveFailed': 'Failed to save token',
     'keybinding.title': 'Keyboard Shortcuts',
     'keybinding.togglePalette': 'Command Palette',
-    'keybinding.openBookmarks': 'Saved Commands',
+    'keybinding.openBookmarks': 'Bookmarks',
     'keybinding.newTab': 'New Tab',
     'keybinding.closeTab': 'Close Tab',
     'keybinding.splitHorizontal': 'Split Horizontal',
@@ -277,9 +281,6 @@ const messages: Record<Locale, Record<string, string>> = {
     'settings.about.title': 'About',
     'settings.about.version': 'Version',
     'settings.about.repository': 'Repository',
-    'settings.about.upToDate': 'Up to date',
-    'settings.about.updateAvailable': 'Update available',
-    'settings.about.download': 'Download',
     'settings.confirmBeforeCloseTab': 'Confirm before closing terminal tab',
     'settings.confirmBeforeCloseTabHint':
       'Show a confirmation dialog before closing a terminal tab',
@@ -288,8 +289,6 @@ const messages: Record<Locale, Record<string, string>> = {
     'confirm.closeTabMessage': 'Closing this session will terminate all running processes. Close',
     'confirm.closeTabConfirm': 'Close',
     'confirm.closeTabCancel': 'Cancel',
-    'settings.about.checkForUpdates': 'Check for updates',
-    'settings.about.checking': 'Checking...',
   },
   zh: {
     'app.settings': '设置',
@@ -349,8 +348,8 @@ const messages: Record<Locale, Record<string, string>> = {
     'terminal.ctxCommand': '命令',
     'terminal.ctxSave': '保存',
     'terminal.ctxNoSelection': '未选中文本',
-    'bookmarks.title': '已保存命令',
-    'bookmarks.empty': '暂无保存的命令',
+    'bookmarks.title': '命令收藏',
+    'bookmarks.empty': '暂无收藏的命令',
     'bookmarks.name': '名称',
     'bookmarks.command': '命令',
     'bookmarks.group': '分组',
@@ -379,8 +378,7 @@ const messages: Record<Locale, Record<string, string>> = {
     'settings.color.ansi': 'ANSI 颜色',
     'settings.color.reset': '重置',
     'settings.tab.general': '通用',
-    'settings.tab.theme': '主题',
-    'settings.tab.text': '文字',
+    'settings.tab.appearance': '外观',
     'settings.tab.keyboard': '键盘',
     'settings.text': '文字设置',
     'settings.text.fontSize': '字号',
@@ -503,9 +501,13 @@ const messages: Record<Locale, Record<string, string>> = {
     'settings.tab.plugins': '插件',
     'settings.plugins.install': '安装插件',
     'settings.plugins.installFile': '从文件安装',
+    'settings.plugins.installFolder': '从文件夹安装',
     'settings.plugins.installed': '已安装插件',
     'settings.plugins.devLink': '开发模式',
     'settings.plugins.devLinkHint': '链接本地插件目录，支持热重载开发',
+    'settings.plugins.devLinkCheckbox': '开发链接（热重载）',
+    'settings.plugins.browseFolder': '浏览...',
+    'settings.plugins.devBadge': 'dev',
     'settings.plugins.load': '加载',
     'settings.plugins.update': '更新',
     'settings.plugins.uninstall': '卸载',
@@ -527,6 +529,7 @@ const messages: Record<Locale, Record<string, string>> = {
     'settings.plugins.noReadme': '暂无详细说明',
     'settings.plugins.back': '返回',
     'settings.plugins.installedBadge': '已安装',
+    'settings.plugins.refresh': '刷新',
     'login.subtitle': '请输入访问令牌以继续',
     'login.placeholder': '访问令牌',
     'login.submit': '登录',
@@ -542,7 +545,7 @@ const messages: Record<Locale, Record<string, string>> = {
     'setup.saveFailed': '保存令牌失败',
     'keybinding.title': '快捷键',
     'keybinding.togglePalette': '命令面板',
-    'keybinding.openBookmarks': '已保存命令',
+    'keybinding.openBookmarks': '命令收藏',
     'keybinding.newTab': '新建标签页',
     'keybinding.closeTab': '关闭标签页',
     'keybinding.splitHorizontal': '左右分屏',
@@ -564,9 +567,6 @@ const messages: Record<Locale, Record<string, string>> = {
     'settings.about.title': '关于',
     'settings.about.version': '版本',
     'settings.about.repository': '仓库',
-    'settings.about.upToDate': '已是最新版本',
-    'settings.about.updateAvailable': '有新版本可用',
-    'settings.about.download': '下载',
     'settings.confirmBeforeCloseTab': '关闭终端 tab 前显示确认',
     'settings.confirmBeforeCloseTabHint': '关闭终端 tab 前弹出确认对话框',
     'settings.behavior': '行为',
@@ -574,8 +574,6 @@ const messages: Record<Locale, Record<string, string>> = {
     'confirm.closeTabMessage': '关闭此会话将终止所有正在运行的进程。仍要关闭',
     'confirm.closeTabConfirm': '关闭',
     'confirm.closeTabCancel': '取消',
-    'settings.about.checkForUpdates': '检查更新',
-    'settings.about.checking': '检查中...',
   },
 }
